# FAST_DEC_LOOP: better decompress performance on ARM64/X86