@parheliamm 2019-06-06
This change is depend on the GCC version.
GCC6 has obvious improvement, but GCC7 no improvements.
GCC4.9 for kernel has downgrade after apply this patch.
So drop it.