@parheliamm 2019-06-07
@Cyan4973 
Please review these 2 patches.
I update the verification case in kernel side, after apply these 2 patches the overall decompress speed is 2% better than 1.8.3

For lzbench, please refer to below result
dev branch with 2 patches.
lz4 1.9.x                 131 MB/s   587 MB/s   302640041  47.60 /sdcard/silesia

dev branch with 16 bytes copy patch
lz4 1.9.x                 126 MB/s   585 MB/s   302640041  47.60 /sdcard/silesia

1.8.3
lz4 1.8.3                 130 MB/s   569 MB/s   302640041  47.60 /sdcard/silesia