@Cyan4973 2019-06-07
Thanks @parheliamm .
I'll look into them.
I'm a bit overwhelmed right now by many other activities, but don't worry, this PR is on my list, I'll get to it.