@parheliamm 2019-06-09
> Thanks @parheliamm .
> I'll look into them.
> I'm a bit overwhelmed right now by many other activities, but don't worry, this PR is on my list, I'll get to it.

Thanks for review my ticket, I will continue updating this ticket for my progress both on kernel and userspace.