@parheliamm 2019-06-11
Hi @Cyan4973:
Here are 2 commits to improve decompress performance on ARM64/X86.
1. FAST_DEC_LOOP: prefer to lerverage 16 bytes copy if offset >= 16 
This patch is only affect ARM64 which has better 16 bytes copy performance compare with 8 bytes.
After this change, the 16 bytes copy invovled if offest >= 16.
2. FAST_DEC_LOOP: Add hints for better code generation. 
This could be generate better instructions for hot path.
3. FAST_DEC_LOOP: skip duplicate offset check if length == ML_MASK 
Reduce duplicate offset check number by 40% to improve the decompress speed.

Test via lzbench:
dev branch with 3 patches.
lz4 1.9.x                 127 MB/s   **586** MB/s   302640041  47.60 /sdcard/silesia
vanilla:
lz4 1.8.3                 128 MB/s   569 MB/s   302640041  47.60 /sdcard/silesia
v1.9.x-348e10(restored FORCE_INLINE)
lz4 1.9.x                 128 MB/s   577 MB/s   302640041  47.60 /sdcard/silesia