@parheliamm 2019-06-14
@Cyan4973 
I've finished the kernel side benchmark test.
ARM64 with GCC build show improved as below:
Patched:
    READ: bw=505MiB/s (529MB/s)
Vanilla:
    READ: bw=521MiB/s (546MB/s)

Unfortunately, Clang for ARM64 in kernel has big downgrade.
So I prefer to only enable FAST_DEC_LOOP with GCC on ARM64 for safety.

Chenxi