@Cyan4973 2019-06-25
Hi @parheliamm ,

I tried this patch on a bunch of platforms, and here are some results.

phone, Qualcomm MSM8996, Android, clang v8.0 : 
calgary : 890 MB/s => 850 MB/s (-4.7%)
silesia : 930 MB/s => 930 MB/s (-0.0%)

laptop, Core i7-8559U, macOS, llvm10 : 
calgary : 4450 MB/s => 4300 MB/s (-3.5%)
silesia : 4310 MB/s => 4240 MB/s (-1.5%)

laptop, Core i7-8559U, macOS, gcc v9.1.0 : 
calgary : 4320 MB/s => 4200 MB/s (-2.8%)
silesia : 4260 MB/s => 4230 MB/s (-0.7%)

desktop, Core i7-9700K, Ubuntu Linux, clang v8.0 : 
calgary : 3570 MB/s => 3490 MB/s (-2.3%)
silesia : 3640 MB/s => 3605 MB/s (-1.0%)

desktop, Core i7-9700K, Ubuntu Linux, gcc v8.3.0 : 
calgary : 3660 MB/s => 3630 MB/s (-0.8%)
silesia : 3760 MB/s => 3780 MB/s (+0.5%)

I understand this patch was mostly intended for `gcc` for `ARM64`, 
unfortunately this combination doesn't compile anymore on my test system, so I could not test it.

For other platforms, as one can see, the results vary depending on platform, while being generally slightly negative.

Unfortunately, it's not a clear positive signal.