@parheliamm 2019-06-30
@Cyan4973:
Sorry for late reply.
I am back to vocation just now.
I checked your test result, I have some questions and comment for your refernce.
First of all, my chances didn't affect clang at all, so below test result is intresting:
phone, Qualcomm MSM8996, Android, clang v8.0 :
calgary : 890 MB/s => 850 MB/s (-4.7%)
silesia : 930 MB/s => 930 MB/s (-0.0%)
desktop, Core i7-9700K, Ubuntu Linux, clang v8.0 :
calgary : 3570 MB/s => 3490 MB/s (-2.3%)
silesia : 3640 MB/s => 3605 MB/s (-1.0%)
Based on my understandings of my changes, the reliable result should be clang test result is same level.
However, your result has downgade.

Your X86 test result, it also intresting, the reason is below:
1. from code analysis, the commit https://github.com/lz4/lz4/pull/725/commits/3d56642c5bc032d382717eafca7aa791e7ec9be8
should reduce lots of condition checks during decompression. 
So the test result should not like this.

So I request you to double check my changes from code level then try the benchmark.
Meanwhile, I will try X86 tomorrow on my desktop to double check as well.

Any feedback would be appreciate and looking forward to reopening this ticket.

Chenxi.