@Cyan4973 2019-07-01
>First of all, my chances didn't affect clang at all, 

I agree that the patch regarding copy-by-16, is not supposed to be applied for `clang`.
But the code is not completely neutral beyond that part. I can see that :

- a few `unlikely()` compiler hints were added, and they will be applied by `clang`. I'm unsure of the impact of these compiler hints. I prefer to keep such statements for situations which are very rare or very common, such as checking for an error case. Here, these situations will happen. It's not clear to me that they are going to be "rare". Hence it might actually hurt performance ? A would be to benchmark such change in isolation.

- There is a new `jmp` instruction, to skip a further test. I'm unable to tell how it translates into binaries, but it could end up changing the code layout in more or less unpredictable ways.

I can indeed make new tests, to check again.
But it seems this topic is also part of your new patch #744 ?